Update store.js - null reference on key
If a null/undefined is passed to the `key` parameter, the call to .replace() throws an error in the module. By initializing the key so it's never null/undefined, we avoid that problem and likely end up returning null/undefined/empty results where the error will only appear in user code (easier to debug too).